sched/fair: Modify throttle time accounting for per-task throttle
Record "cfs_rq->throttled_clock" and "cfs_rq->throttled_clock_self" when
the first task is thorottled on the hierarchy.

To disambiguate dequeue for throttle vs DEQUEUE_SPECIAL, introduce
DEQUEUE_THROTTLE. Start the accounting when DEQUEUE_THROTTLE is set and
"cfs_rq->throttled_clock_*" is 0.

This matches with the stats tracking strategy of per-cfs_rq throttling
that existed before. If all tasks are dequeued after the cfs_rq was
throttled, the "cfs_rq->throttled_clock_*" still reflected the time when
the throttled hierarchy first had a task queued on it and this is no
different.

  [ prateek: Got rid of h_nr_throttled accounting and retained the bare
    minimal changes to keep it consistent with previous behavior. ]

Signed-off-by: Aaron Lu <ziqianlu@bytedance.com>
Signed-off-by: K Prateek Nayak <kprateek.nayak@amd.com>